Relocate files in proyject
