+++ src/App.jsx
@@ -1,5 +1,5 @@
 import "./App.css";
-import MainLayout from "./components/MainLayout/MainLayout";
+import MainLayout from "./components/Layouts/MainLayout/MainLayout";
 
 function App() {
   return <MainLayout />;
