+++ src/components/Layouts/MainLayout/MainLayout.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styles from "./MainLayout.module.css";
 
-import Navbar from "../Navbar/Navbar";
+import Navbar from "../../shared/Navbar/Navbar";
 
 import Boilers from "../../pages/Boilers/Boilers";
 import Customers from "../../pages/Customers/Customers";
@@ -9,6 +9,7 @@ import Reports from "../../pages/Reports/Reports";
 import Schedule from "../../pages/Schedule/Schedule";
 import Settings from "../../pages/Settings/Settings";
 import Technicians from "../../pages/Technicians/Technicians";
+import TechnicianDetail from "../../pages/TechnicianDetail/TechnicianDetail";
 import Home from "../../pages/Home/Home";
 
 import { Switch, Route, Redirect } from "react-router-dom";
@@ -33,7 +34,7 @@ function MainLayout() {
         <Route path="/settings/:id" component={Settings} />
         <Route path="/settings" exact component={Settings} />
 
-        <Route path="/technicians/:id" component={Technicians} />
+        <Route path="/technicians/:id" component={TechnicianDetail} />
         <Route path="/technicians" exact component={Technicians} />
 
         <Route path="/" exact component={Home} />
