+++ src/components/pages/Boilers/Boilers.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import styles from "./Boilers.module.css";
-import Header from "../../components/Header/Header";
+import Header from "../../shared/Header/Header";
 
 function Boilers() {
   return (
