+++ src/components/pages/Home/Home.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import styles from "./Home.module.css";
-import Header from "../../components/Header/Header";
+import Header from "../../shared/Header/Header";
 
 function Home() {
   return (
