+++ src/components/shared/Navbar/Navbar.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styles from "./Navbar.module.css";
 
-import { Switch, Route, NavLink } from "react-router-dom";
+import { NavLink } from "react-router-dom";
 
 function Navbar() {
   return (
