--- src/pages/Technicians/Technicians.module.css
@@ -1,4 +1,0 @@
-/* */
-.container {
-  flex: 1;
-}
